requirements.txt
faiss-cpu==1.7.4 
mistralai==0.0.12
langchain-openai
langchain